renaming, cleaning up and minor improvements

+++ src/main/java/moba/server/datatypes/objects/Train.java
@@ -25,7 +25,7 @@ import moba.server.datatypes.enumerations.TrainType;
 
 // FIXME: Record Train ist abweichend: Hier ist die Id mit hinterlegt!!
 public record Train(
-    int trainId,
+    long trainId,
     int address,
     int speed,
     DrivingDirection drivingDirection,
